refactor(session-store): extract transcript-entry index/patch (file #4)
Lift the transcript-entry index (transcriptEntryIndexes + getTranscriptEntryIndex/
seed/replace/appendSegment + AppendedTranscriptEntryIndexMap + patched-array views)
into transcript-entry-index.ts; session-store re-exports transcriptEntryIndexes.
Pure transcript-snapshot transforms. Net 13/13 + store vitest parity, check 0.
